pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thelighttrading"
version = "0.1.0"
description = "Multi-LLM trading prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "typer",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "requests",
    "pynacl",
    "PyYAML",
]

[project.scripts]
thelighttrading = "thelighttrading.cli.main:app"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
